BUGFIX: Move `CRBehavioralTestsSubjectProvider` to `Neos\ContentRepository\TestSuite`
so that third party packages can use it by requiring the testsuite
